requirements.txt
batch_pystoi==0.0.3
boto3==1.35.56
coverage==7.6.4
matplotlib==3.9.2
myst-parser==4.0.0
numba==0.60.0
numpy==2.0.2
pesq==0.0.4
pre-commit==3.8.0
pytest==8.3.3
python-dotenv==1.0.1
ruff==0.6.4
scipy==1.14.1
soundfile==0.12.1
soxr==0.5.0.post1
Sphinx==8.1.3
sphinx-autoapi==3.3.3
sphinx-rtd-theme==3.0.1
sshfs==2024.9.0
thop==0.1.1.post2209072238
torch==2.5.1
torchaudio==2.5.1
torchcomp==0.2
tqdm==4.66.5
wandb==0.18.0
